requirements.txt
google-adk[a2a]
google-genai
uvicorn
gunicorn
python-dotenv
certifi
